+++ pac4j-core/src/main/java/org/pac4j/core/engine/DefaultLogoutLogic.java
@@ -7,10 +7,7 @@ import org.pac4j.core.context.HttpConstants;
 import org.pac4j.core.context.Pac4jConstants;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.context.session.SessionStore;
-import org.pac4j.core.exception.http.HttpAction;
-import org.pac4j.core.exception.http.NoContentAction;
-import org.pac4j.core.exception.http.RedirectionAction;
-import org.pac4j.core.exception.http.FoundAction;
+import org.pac4j.core.exception.http.*;
 import org.pac4j.core.http.adapter.HttpActionAdapter;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.profile.ProfileManager;
@@ -95,7 +92,7 @@ public class DefaultLogoutLogic<R, C extends WebContext> extends AbstractExcepti
             }
             logger.debug("redirectUrl: {}", redirectUrl);
             if (redirectUrl != null) {
-                action = new FoundAction(redirectUrl);
+                action = RedirectionActionHelper.buildRedirectUrlAction(context, redirectUrl);
             } else {
                 action = NoContentAction.INSTANCE;
             }