+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/GitHubClient.java
@@ -1,7 +1,7 @@
 package org.pac4j.oauth.client;
 
 import com.github.scribejava.apis.GitHubApi;
-import org.pac4j.core.exception.http.FoundAction;
+import org.pac4j.core.exception.http.RedirectionActionHelper;
 import org.pac4j.oauth.profile.github.GitHubProfile;
 import org.pac4j.oauth.profile.github.GitHubProfileDefinition;
 
@@ -35,7 +35,8 @@ public class GitHubClient extends OAuth20Client {
     protected void clientInit() {
         configuration.setApi(GitHubApi.instance());
         configuration.setProfileDefinition(new GitHubProfileDefinition());
-        defaultLogoutActionBuilder((ctx, profile, targetUrl) -> Optional.of(new FoundAction("https://github.com/logout")));
+        defaultLogoutActionBuilder((ctx, profile, targetUrl) ->
+            Optional.of(RedirectionActionHelper.buildRedirectUrlAction(ctx, "https://github.com/logout")));
 
         super.clientInit();
     }