+++ pac4j-oauth/src/main/java/org/pac4j/oauth/redirect/OAuth10RedirectionActionBuilder.java
@@ -5,10 +5,10 @@ import com.github.scribejava.core.model.OAuth1RequestToken;
 import com.github.scribejava.core.oauth.OAuth10aService;
 import org.pac4j.core.client.IndirectClient;
 import org.pac4j.core.exception.http.RedirectionAction;
-import org.pac4j.core.exception.http.FoundAction;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.exception.HttpCommunicationException;
 import org.pac4j.core.exception.TechnicalException;
+import org.pac4j.core.exception.http.RedirectionActionHelper;
 import org.pac4j.core.redirect.RedirectionActionBuilder;
 import org.pac4j.core.util.CommonHelper;
 import org.pac4j.oauth.config.OAuth10Configuration;
@@ -56,7 +56,7 @@ public class OAuth10RedirectionActionBuilder implements RedirectionActionBuilder
             context.getSessionStore().set(context, configuration.getRequestTokenSessionAttributeName(client.getName()), requestToken);
             final String authorizationUrl = service.getAuthorizationUrl(requestToken);
             logger.debug("authorizationUrl: {}", authorizationUrl);
-            return Optional.of(new FoundAction(authorizationUrl));
+            return Optional.of(RedirectionActionHelper.buildRedirectUrlAction(context, authorizationUrl));
 
         } catch (final OAuthException e) {
             throw new TechnicalException(e);